+++ CHANGELOG.md
@@ -5,6 +5,13 @@ Version numbering represents the Swift version, plus a running number representi
 You can also refer to commit logs to get details on what was implemented, fixed and improved.
 
 
+### 3.0.x
+
+- Move `secretInBody` and `customParameters` from `authConfig` to `clientConfig`, where they belong
+- Allow to override default UTF-8 encoding of the _Basic_ authorization header
+- Improvements to embedded authorization
+
+
 ### 3.0.1
 
 - Add Azure flow (thanks @everlof)